Refactor movie API implementation by modularizing routes and CORS middleware
