+++ clase-4/middlewares/cors.js
@@ -1,0 +1,17 @@
+import cors from 'cors'
+
+const ACCEPTED_ORIGINS = [
+  'http://localhost:8080',
+  'http://localhost:1234',
+  'https://myapp.com'
+]
+
+export const corsMiddleware = ({ acceptedOrigins = ACCEPTED_ORIGINS } = {}) => cors({
+  origin: (origin, callback) => {
+    if (!origin || acceptedOrigins.includes(origin)) {
+      return callback(null, true)
+    }
+
+    return callback(new Error('Not allowed by CORS'))
+  }
+})
