+++ clase-4/utils.js
@@ -1,0 +1,4 @@
+import { createRequire } from 'node:module'
+const require = createRequire(import.meta.url)
+
+export const readJSON = (path) => require(path)
